RM:
Since there werde several versions of the templates on SourceForge
and some were never (e.g. Word 2010 and OO/LO were newer than the
LATeX and MAC version). We investigated a bit during NA-PAC 2013
and discovered that the old accelconf page hosted the new version
of all the template files. I downloaded the templates for LATeX
and Word in A4 and Letter format and replaced the ones on
SourceForge.
I assumed that the OpenOffice/LibreOffice versions on SourceForge
are the most recent ones. Since they looked up-to-date I did not
replace them.